refactor(cli): settings doesn't log anymore
